Bug lost step for biding auction.
Bug: at the step X, the agent wants to send a bid for auction N, but
this agent have already sent a action to the server at step X.
The bid is sent to the server at step X+1, but it is no longer useful,
the auction is gone. 
The execution keeps going and the agent starts to send noActions.

It does not fix the problem. 
I've tried to put a wait after the actionId be received, but it did not
solve the problem.
I just let the bug happens and after it, I fix the course of agent's
actions and keep the bug for biding auctions on metrics. Therefore, the
agent does not send noActions.